+++ src/Schedule.h
@@ -5,7 +5,7 @@
  * Defines the internal representation of the schedule for a function
  */
 
-#include "IR.h"
+#include "Expr.h"
 #include <string>
 #include <vector>
